# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.JsonResponse:
    description: JsonResponse is a struct for response JSON message
    properties:
      Action:
        description: Action is a string for response action
        type: string
      MSG:
        description: Message is an interface for response message can be string, models.DeviceType,
          models.Manufacturer, models.Equipment
      Status:
        description: Status is a string for response status
        example: SUCCESS|ERROR
        type: string
    type: object
  models.Manufacturer:
    description: Manufacturer is a struct for manufacturer
    properties:
      id:
        description: ID is an int32 for manufacturer id
        example: 1
        type: integer
      name:
        description: Name is a string for manufacturer name
        example: Apple
        type: string
      status:
        description: Status is a string for manufacturer status either active or inactive
        example: active|inactive
        type: string
    type: object
info:
  contact: {}
  description: This is the API to interact with Equipment database
  termsOfService: http://swagger.io/terms/
  title: Equipment API
  version: "1.0"
paths:
...
  /manufacturer:
    get:
      consumes:
      - application/json
      description: get all manufacturers from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JsonResponse'
            - properties:
                MSG:
                  $ref: '#/definitions/models.Manufacturer'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.JsonResponse'
      summary: get all manufacturers
      tags:
      - manufacturer
    post:
      consumes:
      - application/json
      description: create manufacturer for the database
      parameters:
      - description: Manufacturer Name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.JsonResponse'
      summary: create manufacturer
      tags:
      - manufacturer
      x-order: 4
  /manufacturer/{id}:
    get:
      consumes:
      - application/json
      description: get a manufacturer by ID from the database
      parameters:
      - description: Manufacturer ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JsonResponse'
            - properties:
                MSG:
                  $ref: '#/definitions/models.Manufacturer'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.JsonResponse'
      summary: get a manufacturer by ID
      tags:
      - manufacturer
  /manufacturer/{id}/name:
    patch:
      consumes:
      - application/json
      description: update a manufacturer name by ID from the database
      parameters:
      - description: Manufacturer ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Manufacturer Name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.JsonResponse'
      summary: update a manufacturer name by ID
      tags:
      - manufacturer
  /manufacturer/{id}/status:
    patch:
      consumes:
      - application/json
      description: update a manufacturer status by ID from the database
      parameters:
      - description: Manufacturer ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Manufacturer Status
        enum:
        - active
        - inactive
        in: query
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JsonResponse'
            - properties:
                MSG:
                  $ref: '#/definitions/models.Manufacturer'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.JsonResponse'
      summary: update a manufacturer status by ID
      tags:
      - manufacturer
swagger: "2.0"